// apps/web/types/api/organization-profile.ts
export interface IOrgnizationPofile {
  title?: string;
  contact_email?: string;
  phone_number?: string;
  bio?: string;
  type: string;
  opportunity_types: string[];
  required_skills: string[];
  state?: string;
  area?: string;
  abn?: string;
  website?: string;
  profile_img?: string;
  cover_img?: string;
}